requirements.txt
agate==1.9.1
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
babel==2.17.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
daff==1.4.2
dbt-adapters==1.16.1
dbt-common==1.27.0
dbt-core==1.10.4
dbt-extractor==0.6.0
dbt-postgres==1.9.0
dbt-protos==1.0.339
dbt-semantic-interfaces==0.9.0
deepdiff==7.0.1
fastapi==0.116.1
filelock==3.18.0
fonttools==4.59.0
fsspec==2025.7.0
h11==0.16.0
idna==3.10
importlib_metadata==8.7.0
isodate==0.6.1
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
leather==0.4.0
MarkupSafe==3.0.2
mashumaro==3.14
matplotlib==3.10.3
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
ordered-set==4.1.0
packaging==25.0
pandas==2.3.1
parsedatetime==2.6
pathspec==0.12.1
pillow==11.3.0
protobuf==5.29.5
psutil==7.0.0
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
pyaes==1.6.1
pyasn1==0.6.1
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
rsa==4.9.1
scipy==1.16.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
snowplow-tracker==1.1.0
SQLAlchemy==2.0.41
sqlparse==0.5.3
starlette==0.47.2
sympy==1.14.0
Telethon==1.40.0
text-unidecode==1.3
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.169
ultralytics-thop==2.0.14
urllib3==2.5.0
uvicorn==0.35.0
zipp==3.23.0
